examples/requirements.txt
# Example Python Dependencies
# Install with: pip install -r requirements.txt

psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
